Factor out some of Formula logic.
