+++ nodes/number/formula3.py
@@ -16,7 +16,6 @@
 #
 # ##### END GPL LICENSE BLOCK #####
 
-import ast
 from math import *
 from collections import defaultdict
 
@@ -28,114 +27,7 @@ import io
 from sverchok.node_tree import SverchCustomTreeNode, throttled
 from sverchok.data_structure import updateNode, match_long_repeat, zip_long_repeat
 from sverchok.utils import logging
-from sverchok.utils.script_importhelper import safe_names
-
-class VariableCollector(ast.NodeVisitor):
-    """
-    Visitor class to collect free variable names from the expression.
-    The problem is that one doesn't just select all names from expression:
-    there can be local-only variables.
-
-    For example, in
-
-        [g*g for g in lst]
-
-    only "lst" should be considered as a free variable, "g" should be not,
-    as it is bound by list comprehension scope.
-
-    This implementation is not exactly complete (at least, dictionary comprehensions
-    are not supported yet). But it works for most cases.
-
-    Please refer to ast.NodeVisitor class documentation for general reference.
-    """
-    def __init__(self):
-        self.variables = set()
-        # Stack of local variables
-        # It is not enough to track just a plain set of names,
-        # since one name can be re-introduced in the nested scope
-        self.local_vars = []
-
-    def push(self, local_vars):
-        self.local_vars.append(local_vars)
-
-    def pop(self):
-        return self.local_vars.pop()
-
-    def is_local(self, name):
-        """
-        Check if name is local variable
-        """
-
-        for stack_frame in self.local_vars:
-            if name in stack_frame:
-                return True
-        return False
-
-    def visit_SetComp(self, node):
-        local_vars = set()
-        for generator in node.generators:
-            if isinstance(generator.target, ast.Name):
-                local_vars.add(generator.target.id)
-        self.push(local_vars)
-        self.generic_visit(node)
-        self.pop()
-
-    def visit_ListComp(self, node):
-        local_vars = set()
-        for generator in node.generators:
-            if isinstance(generator.target, ast.Name):
-                local_vars.add(generator.target.id)
-        self.push(local_vars)
-        self.generic_visit(node)
-        self.pop()
-
-    def visit_Lambda(self, node):
-        local_vars = set()
-        arguments = node.args
-        for arg in arguments.args:
-            local_vars.add(arg.id)
-        if arguments.vararg:
-            local_vars.add(arguments.vararg.arg)
-        self.push(local_vars)
-        self.generic_visit(node)
-        self.pop()
-
-    def visit_Name(self, node):
-        name = node.id
-        if not self.is_local(name):
-            self.variables.add(name)
-
-        self.generic_visit(node)
-
-def get_variables(string):
-    """
-    Get set of free variables used by formula
-    """
-    string = string.strip()
-    if not len(string):
-        return set()
-    root = ast.parse(string, mode='eval')
-    visitor = VariableCollector()
-    visitor.visit(root)
-    result = visitor.variables
-    return result.difference(safe_names.keys())
-
-# It could be safer...
-def safe_eval(string, variables):
-    """
-    Evaluate expression, allowing only functions known to be "safe"
-    to be used.
-    """
-    try:
-        env = dict()
-        env.update(safe_names)
-        env.update(variables)
-        env["__builtins__"] = {}
-        root = ast.parse(string, mode='eval')
-        return eval(compile(root, "<expression>", 'eval'), env)
-    except SyntaxError as e:
-        logging.exception(e)
-        raise Exception("Invalid expression syntax: " + str(e))
+from sverchok.utils.modules.eval_formula import get_variables, safe_eval
 
 class SvFormulaNodeMk3(bpy.types.Node, SverchCustomTreeNode):
     """
